Create Fuel-Consumption
hie everyone im Delaine, im new to coding im trying to write this code for a school project on energy in which we are to calculate the amount of fuel consumption over a certain amount distance using a certain speed as well so I've written this code, and im not sure if its okay n compiling well. if u cn help its a C# code. thank you for yo help.